// src/modules/Discounts/actions/discount/discount.types.ts
export enum DiscountActionTypes {
    FETCH_ALL_DISCOUNTS_START = 'FETCH_ALL_DISCOUNTS_START',
    FETCH_DISCOUNT_BY_ID_START = 'FETCH_DISCOUNT_BY_ID_START',
    SAVE_REDEEMED_DISCOUNT_START = 'SAVE_REDEEMED_DISCOUNT_START',
    TOGGLE_QR_CODE_READER_START = 'TOGGLE_QR_CODE_READER_START',
    TOGGLE_SOCIAL_SHARE_START = 'TOGGLE_SOCIAL_SHARE_START',
    FETCH_ALL_DISCOUNTS_COMPLETED = 'FETCH_ALL_DISCOUNTS_COMPLETED',
    FETCH_DISCOUNT_BY_ID_COMPLETED = 'FETCH_DISCOUNT_BY_ID_COMPLETED',
    SAVE_REDEEMED_DISCOUNT_COMPLETED = 'SAVE_REDEEMED_DISCOUNT_COMPLETED',
    TOGGLE_QR_CODE_READER_COMPLETED = 'TOGGLE_QR_CODE_READER_COMPLETED',
    TOGGLE_SOCIAL_SHARE_COMPLETED = 'TOGGLE_SOCIAL_SHARE_COMPLETED'
}